floss: Update volume callback's signature
To better reflect the actual range of the volume.

Bug: 194853979
Tag: #floss
Test: build, deploy, connect and check dbus-monitor
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 char lines

+++ system/gd/rust/linux/service/src/iface_bluetooth_media.rs
@@ -62,12 +62,12 @@ impl IBluetoothMediaCallback for BluetoothMediaCallbackDBus {
     }
 
     #[dbus_method("OnAbsoluteVolumeChanged")]
-    fn on_absolute_volume_changed(&self, volume: i32) {
+    fn on_absolute_volume_changed(&self, volume: u8) {
         dbus_generated!()
     }
 
     #[dbus_method("OnHfpVolumeChanged")]
-    fn on_hfp_volume_changed(&self, volume: u32, addr: String) {
+    fn on_hfp_volume_changed(&self, volume: u8, addr: String) {
         dbus_generated!()
     }
 }
