+++ system/gd/rust/linux/stack/src/bluetooth_media.rs
@@ -70,13 +70,16 @@ pub trait IBluetoothMediaCallback {
     ///
     fn on_absolute_volume_supported_changed(&self, supported: bool);
 
-    ///
-    fn on_absolute_volume_changed(&self, volume: i32);
+    /// Triggered when a Bluetooth device triggers an AVRCP/A2DP volume change
+    /// event. We need to notify audio client to reflect the change on the audio
+    /// stack. The volume should be in the range of 0 to 127.
+    fn on_absolute_volume_changed(&self, volume: u8);
 
     /// Triggered when a Bluetooth device triggers a HFP AT command (AT+VGS) to
     /// notify AG about its speaker volume change. We need to notify audio
-    /// client to reflect the change on the audio stack.
-    fn on_hfp_volume_changed(&self, volume: u32, addr: String);
+    /// client to reflect the change on the audio stack. The volume should be
+    /// in the range of 0 to 15.
+    fn on_hfp_volume_changed(&self, volume: u8, addr: String);
 }
 
 /// Serializable device used in.
@@ -192,7 +195,7 @@ impl BluetoothMedia {
             }
             AvrcpCallbacks::AvrcpAbsoluteVolumeUpdate(volume) => {
                 self.for_all_callbacks(|callback| {
-                    callback.on_absolute_volume_changed(i32::from(volume));
+                    callback.on_absolute_volume_changed(volume);
                 });
             }
         }
