+++ system/gd/rust/topshim/src/profiles/hfp.rs
@@ -86,7 +86,7 @@ pub mod ffi {
     extern "Rust" {
         fn hfp_connection_state_callback(state: u32, addr: RustRawAddress);
         fn hfp_audio_state_callback(state: u32, addr: RustRawAddress);
-        fn hfp_volume_update_callback(volume: u32, addr: RustRawAddress);
+        fn hfp_volume_update_callback(volume: u8, addr: RustRawAddress);
     }
 }
 
@@ -106,7 +106,7 @@ impl Into<RawAddress> for ffi::RustRawAddress {
 pub enum HfpCallbacks {
     ConnectionState(BthfConnectionState, RawAddress),
     AudioState(BthfAudioState, RawAddress),
-    VolumeUpdate(u32, RawAddress),
+    VolumeUpdate(u8, RawAddress),
 }
 
 pub struct HfpCallbacksDispatcher {
@@ -134,7 +134,7 @@ cb_variant!(
 cb_variant!(
     HfpCb,
     hfp_volume_update_callback -> HfpCallbacks::VolumeUpdate,
-    u32, ffi::RustRawAddress -> RawAddress, {
+    u8, ffi::RustRawAddress -> RawAddress, {
         let _1 = _1.into();
     }
 );
